pyproject.toml
[project]
name = "job-apply"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "google-genai>=1.0.0",
    "lxml>=6.0.2",
    "ollama>=0.6.1",
    "openai>=2.8.1",
    "pillow>=10.0.0",
    "playwright>=1.56.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
    "selenium>=4.38.0",
    "webdriver-manager>=4.0.2",
]
